+++ tests/integrations/import/util.rs
@@ -2,11 +2,13 @@
 
 use std::{sync::Arc, thread, time::Duration};
 
+use engine_rocks::RocksEngine;
 use futures::{executor::block_on, stream, SinkExt};
 use grpcio::{ChannelBuilder, Environment, Result, WriteFlags};
 use kvproto::{import_sstpb::*, kvrpcpb::*, tikvpb::*};
 use security::SecurityConfig;
 use test_raftstore::*;
+use test_raftstore_v2::WrapFactory;
 use tikv::config::TikvConfig;
 use tikv_util::HandyRwLock;
 use uuid::Uuid;
@@ -66,6 +68,56 @@ pub fn open_cluster_and_tikv_import_client(
     let import = ImportSstClient::new(ch);
 
     (cluster, ctx, tikv, import)
+}
+
+pub fn open_cluster_and_tikv_import_client_v2(
+    cfg: Option<TikvConfig>,
+    cluster: &mut test_raftstore_v2::Cluster<
+        test_raftstore_v2::ServerCluster<RocksEngine>,
+        RocksEngine,
+    >,
+) -> (Context, TikvClient, ImportSstClient) {
+    let cfg = cfg.unwrap_or_else(|| {
+        let mut config = TikvConfig::default();
+        config.server.addr = "127.0.0.1:0".to_owned();
+        let cleanup_interval = Duration::from_millis(10);
+        config.raft_store.cleanup_import_sst_interval.0 = cleanup_interval;
+        config.server.grpc_concurrency = 1;
+        config
+    });
+    cluster.cfg = Config {
+        tikv: cfg.clone(),
+        prefer_mem: true,
+    };
+    cluster.run();
+
+    let region_id = 1;
+    let leader = cluster.leader_of_region(region_id).unwrap();
+    let epoch = cluster.get_region_epoch(region_id);
+    let mut ctx = Context::default();
+    ctx.set_region_id(region_id);
+    ctx.set_peer(leader);
+    ctx.set_region_epoch(epoch);
+
+    let ch = {
+        let env = Arc::new(Environment::new(1));
+        let node = ctx.get_peer().get_store_id();
+        let builder = ChannelBuilder::new(env)
+            .http2_max_ping_strikes(i32::MAX) // For pings without data from clients.
+            .keepalive_time(cluster.cfg.server.grpc_keepalive_time.into())
+            .keepalive_timeout(cluster.cfg.server.grpc_keepalive_timeout.into());
+
+        if cfg.security != SecurityConfig::default() {
+            let creds = test_util::new_channel_cred();
+            builder.secure_connect(&cluster.sim.rl().get_addr(node), creds)
+        } else {
+            builder.connect(&cluster.sim.rl().get_addr(node))
+        }
+    };
+    let tikv = TikvClient::new(ch.clone());
+    let import = ImportSstClient::new(ch);
+
+    (ctx, tikv, import)
 }
 
 pub fn new_cluster_and_tikv_import_client()
